requirements.txt
mysqlclient
djangorestframework